packages/quantum-language-optimizer/quantum_language_optimizer-0.1.0.tar.gz/quantum_language_optimizer-0.1.0/quantum_language_optimizer/core/quantum_circuit.py
class AbstractQuantumCircuit:
    """
    Abstract quantum circuit class, defining the interface for quantum circuit builders.
    """

    def initialize_state(self, input_data):
        raise NotImplementedError

    def apply_gates(self):
        raise NotImplementedError

    def measure_state(self):
        raise NotImplementedError

    def execute_circuit(self):
        raise NotImplementedError


class SimpleQuantumCircuit(AbstractQuantumCircuit):
    """
    Concrete implementation of a simple quantum circuit, temporarily independent of any quantum computing framework.
    """

    def __init__(self, num_qubits):
        self.num_qubits = num_qubits
        self.state = [0] * num_qubits
        self.gates = []

    def initialize_state(self, input_data):
        """
        Initialize the quantum state.

        Parameters:
        input_data (str): A binary string representing the initial quantum state.
        """
        self.state = [int(bit) for bit in input_data]

    def apply_gates(self):
        """
        Apply quantum gate operations.
        Here, we only record the applied quantum gates and apply them during actual execution.
        """
        self.gates.append('H 0')
        self.gates.append('CX 0 1')
        self.gates.append('RX 3.14/4 2')

    def measure_state(self):
        """
        Measure the quantum state.
        Here, we only record the measurement operation and apply it during actual execution.
        """
        self.gates.append('MEASURE')

    def execute_circuit(self):
        """
        Execute the quantum circuit and return the measurement results.
        Here, we only simulate execution and return a virtual result.
        """
        # Simulate executing the quantum circuit
        result = {'000': 512, '111': 512}  # This is a virtual result
        return result

    def get_circuit(self):
        """
        Return the current description of the quantum circuit.

        Returns:
        str: The current description of the quantum circuit.
        """
        return '\n'.join(self.gates)
